refactor(tests): split ExpandedSurfaceIntegrationTests into 4 focused classes (test-suite-expanded-surface-class-split)
Mirror the IntegrationTests_* pattern from PR #793. Split the 683-line,
17-method ExpandedSurfaceIntegrationTests into three new sibling files
plus a trimmed original:

- ExpandedSurfaceIntegrationTests.ToolContract.cs:
  ExpandedSurfaceIntegrationTests_ToolContract (3 tests + FakeCodeActionService
  + GetPromptText helper + local FindDocumentPath single-arg overload):
  Resources_Return_Machine_Readable_Contracts,
  Prompts_Return_Actionable_Context,
  PreviewCodeAction_Serializes_Service_Response.

- ExpandedSurfaceIntegrationTests.RepoSolutionAnalysis.cs:
  ExpandedSurfaceIntegrationTests_RepoSolutionAnalysis with class-level
  [TestCategory("RepoSolution")] (1 test):
  Reflection_And_Di_Analysis_Run_On_Repo_Solution. Per-method
  [TestCategory("RepoSolution")] removed (redundant under class-level).

- ExpandedSurfaceIntegrationTests.CoverageProcess.cs:
  ExpandedSurfaceIntegrationTests_CoverageProcess with class-level
  [TestCategory("Process")] (2 tests):
  TestCoverageTool_Returns_Structured_Response,
  TestCoverageTool_WhenCoverletMissing_ReturnsStructuredErrorBeforeRunning.
  Per-method [TestCategory("Process")] removed from the first; class-level
  attribute correctly scopes both.

- Trimmed ExpandedSurfaceIntegrationTests.cs retains the 11 general
  shared-workspace tool-contract tests + FindDocumentPath (both overloads)
  + LoadWorkspaceCopyAsync helpers.

Each new class duplicates the [DoNotParallelize] + [TestClass] decoration,
[ClassInitialize] (calling InitializeServices() + LoadSharedSampleWorkspaceAsync),
and [ClassCleanup] (calling DisposeServices()) from the original, sharing
the cached MSBuild workspace via SharedWorkspaceTestBase's path-keyed cache.

No product behavior changes. Pure move; no assertion changes.

Closes: test-suite-expanded-surface-class-split

Validation:
- mcp__roslyn__project_diagnostics on RoslynMcp.Tests: 0 errors / 0 warnings.
- mcp__roslyn__test_run --filter FullyQualifiedName~ExpandedSurfaceIntegrationTests:
  17/17 pass across the 4 resulting classes.
- mcp__roslyn__test_run --filter "TestCategory=RepoSolution&FullyQualifiedName~ExpandedSurface":
  1/1 pass (Reflection_And_Di_Analysis_Run_On_Repo_Solution).
- mcp__roslyn__test_run --filter "TestCategory=Process&FullyQualifiedName~ExpandedSurface":
  2/2 pass (both coverage tests).
- Self-hosted CI runner gates the full release-verify on push.
